Use embedded artifact module directly
