+++ crates/zk-mobile-bench/src/lib.rs
@@ -27,8 +27,8 @@ use world_id_primitives::{
     AuthenticatorPublicKeySet, FieldElement, TREE_DEPTH, authenticator::oprf_query_digest,
 };
 use world_id_proof::{
+    artifacts::embedded,
     circuit_inputs::{NullifierProofCircuitInput, QueryProofCircuitInput},
-    proof,
 };
 
 use fixtures::{first_leaf_merkle_path, generate_rp_fixture};
@@ -52,7 +52,7 @@ fn generate_query_input() -> (
 
     // Load embedded proving material
     let query_material =
-        proof::load_embedded_query_material().expect("failed to load query material");
+        embedded::load_embedded_query_material().expect("failed to load query material");
 
     // Create user keys
     let user_sk = EdDSAPrivateKey::random(&mut rng);
@@ -108,9 +108,9 @@ fn generate_nullifier_input() -> (
 
     // Load embedded proving materials
     let query_material =
-        proof::load_embedded_query_material().expect("failed to load query material");
+        embedded::load_embedded_query_material().expect("failed to load query material");
     let nullifier_material =
-        proof::load_embedded_nullifier_material().expect("failed to load nullifier material");
+        embedded::load_embedded_nullifier_material().expect("failed to load nullifier material");
 
     // Create user keys
     let user_sk = EdDSAPrivateKey::random(&mut rng);
