+++ services/oprf-node/src/auth.rs
@@ -286,7 +286,8 @@ mod tests {
             let mut rng = rand::thread_rng();
 
             let query_material =
-                world_id_proof::load_embedded_query_material().expect("Can load query material");
+                world_id_proof::artifacts::embedded::load_embedded_query_material()
+                    .expect("Can load query material");
 
             let query_blinding_factor = BlindingFactor::rand(&mut rng);
